Require all directory services to provide a realm name.
Add a guid attribute to IDirectoryService and implement it on all
services by generating it from a class-specific GUID and the realm
name.

Generate a GUID for all directory records (if they don't already have
one) by generating it from the service's GUID and the record's type
and short name.

End result is that all principals now have a GUID which is a valid
UUID.

principalUID() is returns the principal's GUID in
DirectoryPrincipalResource.


git-svn-id: http://svn.calendarserver.org/repository/calendarserver/CalendarServer/trunk@761 e27351fd-9f3e-4f54-a53b-843176b1656c