Allow the Memcacher to use other caching backends. There is now a simple dict based cacher that is used for single instance servers when memcached is not available. There is a "null" cacher used when memcached is not available in multi-instance mode (no caching is done).
This allows us to hide the cache availability behind the Memcacher class making the logic a little easier
where ever the class is used. It also allows us to run tests with the full caching logic in place so we
can properly exercise that without the need to have memcached running.

git-svn-id: http://svn.calendarserver.org/repository/calendarserver/CalendarServer/trunk@2502 e27351fd-9f3e-4f54-a53b-843176b1656c